Significant refactor of localization system.
Previously, Yarn scripts stored their localisations on the imported
YarnProgram, and the DialogueRunner requested this information from the
YarnProgram directly. In this change, localized content is stored inside
Localization objects, which are managed by a LocalizedDatabase; the
DialogueRunner now accesses localized content from a LineProvider, which
is a MonoBehaviour that has a reference to a LocalizedDatabase and
produces LocalizedLine objects for delivery to DialogueViewBase
subclasses.

The key types involved are:

LocalizedDatabase: Holds references to Localization objects.

Localization: One per locale. Holds localized strings, assets, and (if
the Addressable Assets package is installed) asset references. In the
editor, maintains a collection of YarnPrograms that it uses to pull
localized content from.

LineProvider: An abstract class that has a method for taking a line ID
and returning a LocalizedLine. This class is subclassed; subclasses can
return a subclass of LocalizedLine if they need to attach additional
data to a LocalizedLine (like AudioClip references).

TextLineProvider: A concrete subclass of LineProvider that returns a
LocalizedLine. Use cases: games without voiceover, games that use their
own audio engine (like FMOD).

AudioLineProvider: A concrete subclass of LineProvider that delivers an
AudioLocalizedLine. Use case: games that use Unity's AudioSource
component.

AddressableAudioLineProvider: A concrete subclass of LineProvider;
identical to AudioLineProvider, but uses Addressable Assets.

LocalizedLine: An object that represents a line, in the user's desired
localization.

AudioLocalizedLine: A subclass of LocalizedLine that has a field for an
AudioClip.

YarnProgram.YarnTranslation: This was previously an independent type;
it's now a class inside YarnProgram. Same behaviour as before
(effectively a key-value pair mapping language names to TextAssets).

YarnImporter now always produces a YarnProgram, even if the Yarn script
has errors - if this is the case the compiled yarn program bytes will be
null.