interrupt and continue now share their actions
Now pressing the continue action (whatever that is configured to be) will call the normal OnViewUserIntentNextLine call on the runner.
The runner is aware of the current line's state and will either call the interrupt or the advance depending on that state.
Now views don't have to do any internal state tracking at all unless they want to for their own needs.