line state is now in the control of views
Lots of changes here that all boil down to the runner now no longer has any care around line presentation and all of that goes into a particular view (or your choice).
This view gets told about user actions and can take it from there.

Input View is no longer a view and will have to be renamed, now its just a front end to a user driven action.
Line View now tracks its state internally and uses that to determine if something is to be an interrupt or not.
Had to stub out lots of things, so much needs to be cleaned up.