Expose line metadata for a Yarn project.
The Yarn compiler is able to parse and expose metadata for each line
(those usually begin with a hashtag '#'). However, this metadata was
not exposed on the Unity layer, and was instead just ignored. This
commit exposes this metadata through the `YarnProject` object.

The metadata is also added as an additional column for each exported
strings file (users will notice a new 'metadata' column in their CSV
files). However, the metadata is currently assumed to be the same for
all localisations, so any localisation will use the same metadata as
the base localisation.