ICONS: Fix No Texture Access editor exception error, NPOT, filesize
I'm not sure which Unity version is started with, but in 2022.2 I was getting an exception when importing YARN files because the icons it uses aren't set to "Readable". This resulted in there being no icons on the files either.

I also adjusted the cropping on the icons so they can be efficient power-of-two sizes, some were 257 on one side instead of 256.
The lossless compression on them has also been improved, reducing file size.