Runner performs more line readiness checks.
Previously depending on the flow into a line it was possible that the readiness check on the line provider was skipped.
This meant that assets for the line weren't loaded in time for the line provider then needing to hand over the line.
Now this check is performed and the runner will wait until the line provider says it is ready.
This can however lead to potentially long delays before a line is ready, not likely tbh, but something to be wary of if you use a lot of very big nodes with loads of assets.