Make TMP resource installation an externally-invoked task
Installing the TMP Essential Resources package appears to hang
Unity 2023 when beginning the tests. This change moves the package
installation to a static method that's intended to be invoked on the
command line by the test runner, to ensure that the TMP Essential
Resources are in the right place before running tests.

unity-editor -batchmode -executeMethod \
  Yarn.Unity.Tests.RuntimeTestUtility.EnsureTMPResourcesAvailableAndExit